+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Upcoming release]
+- Fail CI build if there are uncommitted changes after `npm run build`.
 
 ## [2.5.1]
 - Fixed bug when the quiz can't be finished due to missing complete callback.